refactored bulk claim api tests
