+++ e2e/api/steps_Bulk.js
@@ -6,14 +6,22 @@ const {assert} = chai;
 
 const apiRequest = require('./apiRequest.js');
 const bulkClaimData = require('../fixtures/events/createBulkClaim.js');
+const {waitForFinishedBusinessProcess} = require("./testingSupport");
+const config = require("../config");
+const {addUserCaseMapping} = require("./caseRoleAssignmentHelper");
+const nonProdExpectedEvents = require("../fixtures/ccd/nonProdExpectedEventsLRSpec");
+const claimDataBulk = require("../fixtures/events/createClaimSpecBulk");
+
+let caseId, eventName;
+let caseData = {};
 
 const data = {
-  CREATE_BULK_CLAIM: (mpScenario, interest, customerId, amount, postcodeValidation) =>
-    bulkClaimData.bulkCreateClaimDto(mpScenario, interest, customerId, amount, postcodeValidation),
+  CREATE_BULK_CLAIM: (mpScenario, interest, customerId, amount, postcodeValidation) => bulkClaimData.bulkCreateClaimDto(mpScenario, interest, customerId, amount, postcodeValidation),
+  CREATE_BULK_CLAIM_VIA_CIVILSERVICE: (scenario, pbaV3) => claimDataBulk.createClaimBulk(scenario, pbaV3),
+
 };
 
 module.exports = {
-
   /**
    * Creates a Bulk claim via SDT
    *
@@ -22,7 +30,7 @@ module.exports = {
    * @param interest
    * @return {Promise<void>}
    */
-  /* createClaimFromSDTRequest: async (user, mpScenario, interest) => {
+  /*createClaimFromSDTRequest: async (user, mpScenario, interest) => {
     let createClaimData;
 
     createClaimData = data.CREATE_BULK_CLAIM(mpScenario, interest, '12345678', '87989');
@@ -32,7 +40,26 @@ module.exports = {
     const response_msg = await apiRequest.createBulkClaim('112345', createClaimData);
     assert.equal(response_msg.errorText, 'Unknown User, ');
     assert.equal(response_msg.errorCode, '001');
-  }, */
+  },*/
+
+  createNewClaimWithCaseworkerCivilService: async (user, scenario) => {
+    eventName = 'CREATE_CLAIM_SPEC';
+    caseId = null;
+    caseData = {};
+    let createClaimData  = {};
+
+    createClaimData = data.CREATE_BULK_CLAIM_VIA_CIVILSERVICE(scenario);
+    await apiRequest.setupTokens(user);
+    await assertCaseworkerSubmittedNewClaim('PENDING_CASE_ISSUED', createClaimData);
+    await waitForFinishedBusinessProcess(caseId);
+    console.log('Bulk claim created with case id: ' + caseId);
+    if (scenario === 'ONE_V_ONE') {
+      await assertCorrectEventsAreAvailableToUserBulkClaims(config.bulkClaimSystemUser, 'AWAITING_RESPONDENT_ACKNOWLEDGEMENT');
+    } else {
+      // one v two/multiparty continuing online not currently supported for LiPs
+      await assertCorrectEventsAreAvailableToUserBulkClaims(config.bulkClaimSystemUser, 'PROCEEDS_IN_HERITAGE_SYSTEM');
+    }
+  },
 
   createClaimFromSDTRequestForPostCodeNegative: async (user, mpScenario, interest) => {
     let createClaimData;
@@ -66,4 +93,26 @@ module.exports = {
     assert.equal(response_msg.errorText, 'Bad data, Request already processed');
     assert.equal(response_msg.errorCode, '000');
   },
+};
+
+const assertCaseworkerSubmittedNewClaim = async (expectedState, caseData) => {
+  const response = await apiRequest.submitNewClaimAsCaseworker(eventName, caseData);
+  const responseBody = await response.json();
+  assert.equal(response.status, 201);
+  assert.equal(responseBody.state, expectedState);
+
+  if (eventName === 'CREATE_CLAIM_SPEC') {
+    caseId = responseBody.id;
+    await addUserCaseMapping(caseId, config.applicantSolicitorUser);
+    console.log('Case created: ' + caseId);
+  }
+};
+
+const assertCorrectEventsAreAvailableToUserBulkClaims = async (user, state) => {
+  console.log(`Bulk claim: Asserting user ${user.type} in env ${config.runningEnv} has correct permissions`);
+  const caseForDisplay = await apiRequest.fetchCaseForDisplay(user, caseId);
+  if (['preview', 'demo'].includes(config.runningEnv)) {
+    expect(caseForDisplay.triggers).to.deep.include.members(nonProdExpectedEvents[user.type][state],
+      'Unexpected events for state ' + state + ' and user type ' + user.type);
+  }
 };
